% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{add_fraction_of_identity}
\alias{add_fraction_of_identity}
\title{Adds results from a particular dimension reduction technique to the object.}
\usage{
add_fraction_of_identity(object, exprs_values = "counts", features = NULL,
  slot = "fraction_of_identity", ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used to calculate the fraction of identity.}

\item{features}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use for calculation. The default of NULL will use all features.}

\item{slot}{Determines which entry of the \code{reducedDims} slot to use for reduced embedding.}

\item{...}{Additional parameters passed to \code{calcFractionOfIdentity}.}
}
\value{
A SingleCellExperiment object with modified \code{reducedDims} slot.
}
\description{
Adds results from a particular dimension reduction technique to the object.
}
