% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{add_heterogeneity}
\alias{add_heterogeneity}
\title{Adds a heterogeneity statistic to the object.}
\usage{
add_heterogeneity(object, exprs_values = "counts",
  column = ".heterogeneity", ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used calculate heterogeneous genes.}

\item{column}{Determines the column name of the \code{\link{rowData}} slot to store results to.}

\item{...}{Additional parameters passed to \code{\link{heterogeneity}}}
}
\value{
A SingleCellExperiment object with modified \code{rowData} slot.
}
\description{
Adds a heterogeneity statistic to the object.
}
