% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_index.R
\name{add_transition_index}
\alias{add_transition_index}
\title{Adds the cirtical transition index to the object.}
\usage{
add_transition_index(object, column, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{column}{Determines the column name of the \code{colData} slot to store results to.}

\item{...}{Additional parameters passed to \code{computeIC}.}
}
\value{
A SingleCellExperiment object with modified \code{colData} slot.
}
\description{
Adds the cirtical transition index to the object.
}
