% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AnnotateCell}
\alias{AnnotateCell}
\title{Visualize Summarized Loadings}
\usage{
AnnotateCell(
  seurat_object,
  cell_names,
  group_ident = "cell_types",
  dimplot_ident = "orig.ident"
)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell data.}

\item{cell_names}{A character vector specifying the cell names to annotate.}

\item{dimplot_ident}{A string specifying the metadata variable to group by in the DimPlot.}
}
\value{
A Seurat object with cell names annotated.
}
\description{
This function visualizes the aggregated loadings across principal components.
}
