% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ApplyDoubletFinder}
\alias{ApplyDoubletFinder}
\title{Apply DoubletFinder}
\usage{
ApplyDoubletFinder(seurat_object)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell data.}
}
\value{
A modified Seurat object with doublet information.
}
\description{
This function applies DoubletFinder to identify potential doublets in the single-cell data.
}
