% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{bayesian_blocks}
\alias{bayesian_blocks}
\title{Wrapper for astroML.density_estimation.bayesian_blocks}
\usage{
bayesian_blocks(t, x = NULL, sigma = NULL, fitness = "events", ...)
}
\arguments{
\item{t}{A numeric}

\item{x}{A numeric}

\item{sigma}{Data error}

\item{fitness}{Fitness function}

\item{...}{Other parameters}
}
\value{
array containing the (N+1) bin edges
}
\description{
Wrapper for astroML.density_estimation.bayesian_blocks
}

