% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_index.R
\name{boot.ic}
\alias{boot.ic}
\title{Run non-parametric bootstrapping for IC values and groups}
\usage{
boot.ic(data, groups, n = 20, R = 999, ...)
}
\arguments{
\item{data}{The expression matrix (genes-by-cells)}

\item{groups}{A vector giving group information (will be coerced to factor)}

\item{n}{The subsample size for non-parametric bootstrapping}

\item{R}{The bootstrapping size}
}
\value{
A matrix with all bootstrap samples per grouping level (R rows, length(levels(group)) columns)
}
\description{
Run non-parametric bootstrapping for IC values and groups
}
