% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_index.R
\name{boot_subsample}
\alias{boot_subsample}
\title{Helper function for non-parametric subsampling}
\usage{
boot_subsample(x, subsample_size)
}
\arguments{
\item{x}{The data vector (or matrix)}

\item{subsample_size}{The size of the subsample}
}
\value{
A matrix in the same dimensions as x, but subsampled to subsample_size rows.
}
\description{
Helper function for non-parametric subsampling
}
