% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_index.R
\name{bootstrap_transition_index}
\alias{bootstrap_transition_index}
\title{Run non-parametric bootstrapping for critical transition index values and a grouping variable}
\usage{
bootstrap_transition_index(object, exprs_values, grouping,
  features = NULL, n = 20, R = 999, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used to calculate IC.}

\item{grouping}{A single character indicating the column of \code{colData} to use as grouping, or a factor of \code{length > 1}.}

\item{features}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use for calculation. The default of NULL will use all features.}

\item{n}{The subsample size for non-parametric bootstrapping.}

\item{R}{The bootstrapping size.}

\item{...}{Additional parameters passed to \code{...}.}
}
\value{
A matrix with all bootstrap samples per grouping level (\code{R} rows, \code{length(levels(group))} columns)
}
\description{
Run non-parametric bootstrapping for critical transition index values and a grouping variable
}
