% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{bw.inequality.fun}
\alias{bw.inequality.fun}
\title{Calculates the inequality between and within groups}
\usage{
bw.inequality.fun(x, g)
}
\arguments{
\item{x}{The numeric vector of values}

\item{g}{The grouping factor}
}
\value{
The inequality measure for x given g
}
\description{
Calculates the inequality between and within groups
}
