% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calc_dropout}
\alias{calc_dropout}
\title{Caclulates drop-outs based on burst size and frequency}
\usage{
calc_dropout(size, frequency, location = 1.2, scale = 1.5)
}
\arguments{
\item{size}{The genes burst size}

\item{frequency}{The genes burst frequency}

\item{location}{The location parameter of the bivariate logistic distribution}

\item{scale}{The scale parameter of the bivariate logistic distribution}
}
\value{
The gene-specific drop-out rate
}
\description{
Caclulates drop-outs based on burst size and frequency
}

