% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{calcFNWeight}
\alias{calcFNWeight}
\title{Calculate a weight matrix using false-negative curves}
\usage{
calcFNWeight(data, condition = NULL, means = NULL, expression_cutoff = 10,
  supress.plot = T)
}
\arguments{
\item{data}{Expression data matrix}

\item{condition}{A factor describing different conditions}

\item{means}{A matrix containing mean expression values, e.g. from bulk measurements}

\item{expression_cutoff}{A numerical value}

\item{supress.plot}{Whether to supress plotting (default TRUE)}
}
\description{
Calculate a weight matrix using false-negative curves
}

