% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{calcFractionOfIdentity}
\alias{calcFractionOfIdentity}
\title{Orders cells along a pseudotime between two or more known states.
This method implements the fraction of identity approach of Treutlein et. al. (2016)}
\usage{
calcFractionOfIdentity(data, states)
}
\arguments{
\item{data}{The single-cell expression matrix}

\item{states}{A matrix containing the expression values for known states}
}
\value{
A list with the following elements:
identity - a matrix with one column per state, giving a numerical "identity" of each cell with that state, as well as the error per cell,
order - a matrix with one column per state, giving the decreasing order of the cells identity with that state.
}
\description{
Orders cells along a pseudotime between two or more known states.
This method implements the fraction of identity approach of Treutlein et. al. (2016)
}
