% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{calcMeansPerCondition}
\alias{calcMeansPerCondition}
\title{Calculates means per condition}
\usage{
calcMeansPerCondition(data, condition, log = T)
}
\arguments{
\item{data}{Expression data matrix}

\item{condition}{A factor describing different conditions}

\item{log}{If the results should be returned in log scale.}
}
\description{
Calculates means per condition
}
