% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CalcMetaGene}
\alias{CalcMetaGene}
\title{Calculate MetaGene Expression}
\usage{
CalcMetaGene(seurat_object, gene_list)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell data.}

\item{gene_list}{A character vector specifying the gene list to calculate the average expression.}
}
\value{
A numeric vector representing the average expression of the gene list.
}
\description{
This function calculates the average expression of a gene list across cells.
}
