% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calcSOM}
\alias{calcSOM}
\title{Calculates a (weighted) self-organizing map}
\usage{
calcSOM(data, train = NULL, weights = NULL, num_epochs = 200,
  resolution = 1, seed = NULL, init = NULL, init.FUN = map.init,
  parallel = F)
}
\arguments{
\item{data}{A numerical matrix or data.frame.}

\item{train}{Row indices of training data.}

\item{weights}{Additonal weights per row}

\item{num_epochs}{How long the training should last.}

\item{resolution}{A multiplier for the map size}

\item{seed}{A random seed (default NULL).}

\item{init}{A list with initialization parameters.}

\item{parallel}{Indicates if the parallel implementation should be used.}
}
\value{
A som object.
}
\description{
Calculates a (weighted) self-organizing map
}
