% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{calculate_qc_map}
\alias{calculate_qc_map}
\title{Calculates a 2D QC map of cells based on quality features}
\usage{
calculate_qc_map(object, features, slot = "qcmap", ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{features}{A character vector with column names form \code{colData(object)} that are used to calculate the map.}

\item{slot}{Determines which entry of the \code{reducedDims} slot to use for QC embedding.}

\item{...}{Additional parameters passed onto \code{\link[Rtsne]{Rtsne}}.}
}
\value{
A SingleCellExperiment object with modified \code{reducedDims} slot.
}
\description{
Calculates a 2D QC map of cells based on quality features
}
