% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CalculateAUCell}
\alias{CalculateAUCell}
\title{Calculate AUCell scores and add to Seurat object}
\usage{
CalculateAUCell(seurat_object, gene_list, col.name)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell data.}

\item{gene_list}{A character vector specifying the gene list to calculate the module score.}

\item{col.name}{A string specifying the column name to store the AUCell scores.}
}
\value{
A Seurat object with AUCell scores added as metadata.
}
\description{
This function visualizes the aggregated loadings across principal components.
}
