% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calculateClusterCenter}
\alias{calculateClusterCenter}
\title{Calculates cluster centers based on a probability threshold.}
\usage{
calculateClusterCenter(c, min.probability = 0.4)
}
\arguments{
\item{c}{The clustering}

\item{min.probability}{The minimum probability}
}
\value{
The cluster centers
}
\description{
Calculates cluster centers based on a probability threshold.
}
