% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calculateSilhouette}
\alias{calculateSilhouette}
\title{Calculates silhouette values for each row in data}
\usage{
calculateSilhouette(data, clusters, dims_to_use = 1:2, ...)
}
\arguments{
\item{data}{A data matrix that can be used to calculate distances between rows}

\item{clusters}{A vector indicating cluster membership}

\item{dims_to_use}{Specifies which dimensions to use for calculating the distance between observations}

\item{...}{Other arguments passed to the \code{dist} function.}
}
\value{
A vectr of silhouette values.
}
\description{
Calculates silhouette values for each row in data
}
\details{
Implementation as in https://en.wikipedia.org/wiki/Silhouette_(clustering)
}
