% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ClusterComp}
\alias{ClusterComp}
\title{Cluster Composition Analysis}
\usage{
ClusterComp(
  seurat_object,
  meta_data = "orig.ident",
  idents = NULL,
  cluster_name = "seurat_clusters",
  reduced = FALSE
)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell data.}

\item{meta_data}{A string specifying the metadata variable for original identities.}

\item{idents}{A vector specifying the cluster identities to include in the analysis.}

\item{cluster_name}{A string specifying the name of the cluster identity column in the Seurat object.}

\item{reduced}{A logical specifying whether each orig.ident should be reduced to the smallest cell count.}
}
\value{
A plot showing the cluster composition in terms of original identities.
}
\description{
This function calculates the composition of each cluster in terms of original identities.
}
