% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{colorAMap}
\alias{colorAMap}
\title{A function to color observations on a map based on values of a vector.}
\usage{
colorAMap(data, colour_by, shape_by = NULL,
  palette = RColorBrewer::brewer.pal(8, "Dark2"), na.colour = "darkgrey",
  na.shape = 4, na.cex = 1, xlab = "Dimension 1", ylab = "Dimension 2",
  pch = rep(16, nrow(data)), cex = rep(1, nrow(data)), ...)
}
\arguments{
\item{data}{The data matrix with observation coordinates in the first two columns.}

\item{colour_by}{Either a numerical vector on which the coloring is based, or a numeric or character of length 1 indicating the column of the data matrix from which to take the values, or a factor.}

\item{shape_by}{An optional factor on which point shape is based.}

\item{palette}{The color palette.}

\item{na.colour}{The color of \code{NA}s in \code{colour_by}.}

\item{na.shape}{The shape of \code{NA}s in \code{shape_by}}

\item{na.cex}{The cex of \code{NA} observations.}

\item{xlab, ylab}{Labels of x and y axis.}

\item{pch}{Default point shape.}

\item{cex}{Default point size.}

\item{...}{Other parameters passed to lattice::xyplot}
}
\value{
A lattice xyplot. In case \code{colour_by} or \code{shape_by} is a factor, a legend is drawn.
}
\description{
A function to color observations on a map based on values of a vector.
}
