% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{colortSNE}
\alias{colortSNE}
\title{A function to color a tSNE map based on a vector.}
\usage{
colortSNE(tsne, data, use.pal = "RdBu")
}
\arguments{
\item{tsne}{The data matrix from a tSNE dimension reduction call}

\item{data}{A numerical vector on which the coloring is based}

\item{use.pal}{The RColorBrewer color palette}
}
\description{
A function to color a tSNE map based on a vector.
}

