% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_index.R
\name{compute_transition_index}
\alias{compute_transition_index}
\title{Calculate the critical transition index for cells in a SingleCellExperiment object.
The critical transition index is described in detail in http://www.pnas.org/content/114/9/2271.long.}
\usage{
compute_transition_index(object, exprs_values, features = NULL,
  pval = 0.01)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used to calculate the IC.}

\item{features}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use for calculation. The default of NULL will use all features.}

\item{pval}{The p-value cutoff for correlation values to use in calculations}
}
\value{
A list with the IC value (ic), as well as the used correlation values (within (GxG) / between (CxC))
}
\description{
Calculate the critical transition index for cells in a SingleCellExperiment object.
The critical transition index is described in detail in http://www.pnas.org/content/114/9/2271.long.
}
