% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{condprop}
\alias{condprop}
\title{Calculates the conditional propability p(x|given) for two vectors}
\usage{
condprop(x, given)
}
\arguments{
\item{x}{The first vector}

\item{given}{The second vector}
}
\value{
A vector with conditional propabilities.
}
\description{
Calculates the conditional propability p(x|given) for two vectors
}

