% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{construct_filters}
\alias{construct_filters}
\title{Conveneince function to construct a list of filter function(s) from a named list}
\usage{
construct_filters(flist)
}
\arguments{
\item{flist}{A named list with gene filters. Names should be callable functions, list items should be named paramters.}
}
\value{
A list.
}
\description{
Conveneince function to construct a list of filter function(s) from a named list
}
