% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CreateSeuratList}
\alias{CreateSeuratList}
\title{Create Merged Seurat Object}
\usage{
CreateSeuratList(file_paths)
}
\arguments{
\item{file_paths}{A character vector specifying the file paths of the single-cell data.}
}
\value{
A merged Seurat object.
}
\description{
This function creates a merged Seurat object from multiple samples.
}
