% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{createSingleCellExperiment}
\alias{createSingleCellExperiment}
\title{Creates a proper SingleCellExperiment object from expression tables and metadata}
\usage{
createSingleCellExperiment(data.files, metadata.files = NULL,
  cell.identifier = "Sample", gene.identifier = "gene_id",
  annotate.genes = F, biomart.dataset = "mmusculus_gene_ensembl",
  biomart.filter = "ensembl_gene_id", biomart.fields = c("ensembl_gene_id",
  "external_gene_name", "gene_biotype", "description"))
}
\arguments{
\item{data.files}{A named list with file paths to tab-separated expression matrices}

\item{metadata.files}{A list with file paths to tab-separated metadata matrices}

\item{cell.identifier}{The column name of cell identifiers in metadata.files (e.g. Sample, Cell)}

\item{gene.identifier}{The column name of gene identifieres in data.files (e.g. gene_id)}

\item{annotate.genes}{Boolean, whether or not bioMart should be used to annotate genes}

\item{biomart.dataset}{The dataset that is used to fetch annotations (e.g. mmusculus_gene_ensembl)}

\item{biomart.filter}{The filter that is used to match annotations to genes (e.g. ensembl_gene_id, if row.names are Ensembl Gene IDs)}

\item{biomart.fields}{A character vector of field names to fetch}
}
\value{
A SingleCellExperiment object.
}
\description{
Creates a proper SingleCellExperiment object from expression tables and metadata
}
