% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{cv2.fun}
\alias{cv2.fun}
\title{Calculates the coefficient of variation}
\usage{
cv2.fun(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
The coefficient of variation for x
}
\description{
Calculates the coefficient of variation
}
