% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{discover_element}
\alias{discover_element}
\title{Convenience function to discover elements in \code{what} in a vector \code{where} using a function \code{FUN}.}
\usage{
discover_element(object, what, where = c("rownames", "colnames",
  "rowData", "colData"), FUN = c("grepl", "is_in"), column = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{what}{A character vector/character with what to find}

\item{where}{Where to search the object}

\item{FUN}{Function name used to perform the search}

\item{column}{Column name to search in (if rowData or colData should be searched).}
}
\value{
A list with TRUE/FALSE
}
\description{
Convenience function to discover elements in \code{what} in a vector \code{where} using a function \code{FUN}.
}
