% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{discretize.bb}
\alias{discretize.bb}
\title{Discretize a matrix using the bayesian blocks algorithm}
\usage{
discretize.bb(X, ...)
}
\arguments{
\item{X}{Numerical matrix to be discretized column-wise.}

\item{...}{Additional parameters passed to bayesian_blocks}
}
\value{
A discretized matrix
}
\description{
Discretize a matrix using the bayesian blocks algorithm
}
