% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{dm}
\alias{dm}
\title{Calculate the distance to the median}
\usage{
dm(data, ...)
}
\arguments{
\item{data}{A numerical matrix or data.frame.}

\item{window}{The window size to calculate the summary statistic in.}

\item{normalize}{Whether or not to normalize to the standard deviation.}

\item{func}{The function to calculate the summary statistic within windows.}
}
\value{
A vector with the distances to the median.
}
\description{
Calculate the distance to the median
}
