% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsampling of a count matrix to a target size}
\usage{
downsample(counts, totalReads, mean = 1, sd = 0.1,
  lowerDetectionLimit = 10)
}
\arguments{
\item{counts}{The count matrix}

\item{totalReads}{The library size vector (in millions)}

\item{mean}{The mean target library size (in millions)}

\item{sd}{The standard deviation around the mean, used to create target library sizes}

\item{lowerDetectionLimit}{Number of counts considered as lower detection limit}
}
\value{
A list with the following components
  expressed:
}
\description{
Downsampling of a count matrix to a target size
}
