% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{dropout.fun}
\alias{dropout.fun}
\title{Calculates the drouput percentage}
\usage{
dropout.fun(x, limit = 0)
}
\arguments{
\item{x}{The vector of values}

\item{limit}{The detection limit}
}
\value{
The percentage of dropouts
}
\description{
Calculates the drouput percentage
}
