% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{EpOverA}
\alias{EpOverA}
\title{EpOverA returns a filter function with bindings for A. This function evaluates to TRUE if the aggregate of the arguments elements are larger than A.}
\usage{
EpOverA(A = 4, na.rm = TRUE)
}
\arguments{
\item{A}{The value you want to exceed.}

\item{na.rm}{Whether NAs should be removed.}
}
\value{
A function with bindings for A.
}
\description{
EpOverA returns a filter function with bindings for A. This function evaluates to TRUE if the aggregate of the arguments elements are larger than A.
}
