% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{fano.fun}
\alias{fano.fun}
\title{Calculates the Fano factor}
\usage{
fano.fun(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
The index of dispersion for x
}
\description{
Calculates the Fano factor
}
