% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{FanoOverA}
\alias{FanoOverA}
\title{FanoOverA returns a filter function with bindings for A. This function evaluates to TRUE if the Fano factor of the argument is larger than A.}
\usage{
FanoOverA(A = 2, na.rm = TRUE)
}
\arguments{
\item{A}{The value that should be exceeded.}

\item{na.rm}{Whether NAs should be removed.}
}
\value{
A function with bindings for A.
}
\description{
FanoOverA returns a filter function with bindings for A. This function evaluates to TRUE if the Fano factor of the argument is larger than A.
}
\details{
See Weinreb et.al., https://doi.org/10.1093/bioinformatics/btx792
}
