% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_features}
\alias{filter_features}
\title{Conveinience function to perform feature filtering using \code{genefilter}.}
\usage{
filter_features(object, flist, exprs_values = "counts", tolerate = 0)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{flist}{A named list with gene filters. Names should be callable functions, list items should be named paramters.}

\item{exprs_values}{String indicating which assay contains the data that should be used for filtering.}

\item{tolerate}{A number indicating how many failed filters should be toleated.}
}
\value{
A SingleCellExperiment object with features passing the filter(s).
}
\description{
Conveinience function to perform feature filtering using \code{genefilter}.
}
