% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_genes}
\alias{filter_genes}
\title{Conveinience function to perform gene filtering using \code{genefilter}.}
\usage{
filter_genes(object, flist, exprs_values = "counts")
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{flist}{A named list with gene filters. Names should be callable functions, list items should be named paramters.}

\item{exprs_values}{String indicating which assay contains the data that should be used for filtering.}
}
\value{
A SingleCellExperiment object with features passing the filter.
}
\description{
Conveinience function to perform gene filtering using \code{genefilter}.
}
