% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FilterDegs}
\alias{FilterDegs}
\title{Filter Differentially Expressed Genes}
\usage{
FilterDegs(markers, seurat_object, n_genes = 6)
}
\arguments{
\item{markers}{A data frame containing marker genes and their associated statistics.}

\item{seurat_object}{A Seurat object containing single-cell data.}

\item{n_genes}{An integer specifying the number of top genes to select per cluster.}
}
\value{
A list of differentially expressed genes per cluster.
}
\description{
This function filters differentially expressed genes based on average log2 fold change and cluster information.
}
