% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{generate_mutationhash}
\alias{generate_mutationhash}
\title{Generate a mutationhash for a vector of barcodes.}
\usage{
generate_mutationhash(barcodes, edit_distance = 1)
}
\arguments{
\item{barcodes}{Character vector of barcodes}

\item{edit_distance}{Maximum allowed edit distance}
}
\value{
A data.frame with two columns: A column containing barcodes and a column containing possible substituted barcodes.
}
\description{
Generate a mutationhash for a vector of barcodes.
}
