% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getSOMgenes}
\alias{getSOMgenes}
\title{Extracts the rows of the data matrix underlying each SOM node.}
\usage{
getSOMgenes(som, clustering = NULL)
}
\arguments{
\item{som}{The SOM object.}

\item{clustering}{An optional clustering of SOM nodes.}
}
\value{
Row indices or data row names for each node or each cluster as a list.
}
\description{
Extracts the rows of the data matrix underlying each SOM node.
}
