% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{gini.fun}
\alias{gini.fun}
\title{Calculates the gini index}
\usage{
gini.fun(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
The gini index for x
}
\description{
Calculates the gini index
}
