% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GO_KEGG}
\alias{GO_KEGG}
\title{Gene Ontology and KEGG Pathway Analysis}
\usage{
GO_KEGG(seurat_object)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell data.}
}
\value{
A list of dot plots showing the enriched gene ontology and KEGG pathways.
}
\description{
This function performs gene ontology and KEGG pathway analysis on differentially expressed genes.
}
