% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{hdbscan}
\alias{hdbscan}
\title{A wrapper for running HDBSCAN in python}
\usage{
hdbscan(data, min.cluster.size = 5, min.samples = NULL, kwargs = NULL,
  return.data = T)
}
\arguments{
\item{data}{Matrix-like data to be clustered}

\item{min.cluster.size}{The minimum size of clusters}

\item{min.samples}{The number of samples in a neighbourhood for a point to be considered a core point.}

\item{kwargs}{A named list of parameters passed to HDBSCAN}
}
\value{
the clustered data
}
\description{
A wrapper for running HDBSCAN in python
}

