% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{heterogeneity}
\alias{heterogeneity}
\title{Calculates various heterogeneity statistics}
\usage{
heterogeneity(data, statistic = c("cv", "dropout", "gini", "inequality",
  "bwt", "mean"), normalization = c("none", "bins", "windows"),
  order_by = log2(Matrix::rowMeans(data/10, na.rm = T) + 1), groups = NULL,
  ...)
}
\arguments{
\item{data}{A numerical matrix}

\item{statistic}{The heterogeneity measure to calculate. Can be \code{cv} for the coefficient of variation, \code{dropout} for the percentage of dropouts, \code{gini} for the gini index or \code{inequality} for Theil's T statistic.}

\item{normalization}{The normalization method. Can be \code{none} for no normalization, \code{bins} to normalize in bins calculated from the mean or \code{windows} to normalize within rolling windows.}

\item{order_by}{A numeric vector by which the calculated statistic should be ordered in case of normalization.}

\item{groups}{An optional grouping factor.}

\item{...}{Additional arguments passed to the normalization functions.}
}
\value{
The statistics value for each observation.
}
\description{
Calculates various heterogeneity statistics
}
