% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{hex_corner}
\alias{hex_corner}
\title{Calculates corners from a hexagon center.}
\usage{
hex_corner(x, y, size)
}
\arguments{
\item{x}{The hexagon centers x coordinate}

\item{y}{The hexagon centers y coordinate}

\item{size}{The size of the hexagon as height/2.}
}
\value{
A dataframe with x and y coordinates.
}
\description{
Calculates corners from a hexagon center.
}
