% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Hexagon}
\alias{Hexagon}
\title{Creates a polygon}
\usage{
Hexagon(x, y, unitcell = 1, col = col)
}
\arguments{
\item{x}{The polygons x}

\item{y}{The polygons y}

\item{unitcell}{Scaling}

\item{col}{The color of the polygon}
}
\value{
A colored polygon
}
\description{
Creates a polygon
}

