% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{hexcoords2}
\alias{hexcoords2}
\title{Calculates hexagon coordinates from hexagon centers.}
\usage{
hexcoords2(x, y, height = NULL, width = sqrt(3)/2 * height)
}
\arguments{
\item{x}{The x coordinates of the centers}

\item{y}{The y coordinates of the centers}

\item{height}{The hexagons height. Precedes width if given.}

\item{width}{The hexagons width. Will be calculated from height by default.}
}
\value{
A dataframe with the hexagons x and y coordinates, as well a group variable indicating individual hexagons.
}
\description{
Calculates hexagon coordinates from hexagon centers.
}
