% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{hvg.plot}
\alias{hvg.plot}
\title{Draws a smoothScatter-Plot with hvg fit line and highlighted variable genes.}
\usage{
hvg.plot(data, hvg.fit, n = 500)
}
\arguments{
\item{data}{The normalized count table.}

\item{hvg.fit}{A list result from the hvg() function.}

\item{n}{Number of highly variable genes to highlight.}
}
\description{
Draws a smoothScatter-Plot with hvg fit line and highlighted variable genes.
}
