% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{hvg}
\alias{hvg}
\title{Highly variable genes by Brennecke et. al.}
\usage{
hvg(data, min.cv2 = 0.3, mean.quantile = 0.95, padj.method = "fdr")
}
\arguments{
\item{data}{A normalized count table.}

\item{min.cv2}{Minimum coefficient of variation for genes used in fit.}

\item{mean.quantile}{The quantile that should be used to define a minimum mean for genes used in fit.}

\item{padj.method}{Method used in adjusting p-values.}

\item{windsorize}{Whether windsorization should be applied to data}
}
\value{
A list with the fit coefficients, the ordered variance and corresponding adjusted p-values.
}
\description{
Highly variable genes by Brennecke et. al.
}
