% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{inequality.fun}
\alias{inequality.fun}
\title{Calculates the inequality of a numeric vector}
\usage{
inequality.fun(x)
}
\arguments{
\item{c}{The numeric vector of values}
}
\value{
The inequality measure for x
}
\description{
Calculates the inequality of a numeric vector
}
