% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{isomds}
\alias{isomds}
\title{Conveinience function to run isoMDS on a SingleCellExperiment object.}
\usage{
isomds(object, exprs_values, features = NULL, method = "spearman", ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used to perform isoMDS.}

\item{features}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use for isoMDS. The default of NULL will use all features.}

\item{method}{A character string specifying the method to be used to calulate a dissimilarity structure using WGCNA::cor.}

\item{...}{Additional parameters to be passed on to MASS::isoMDS.}
}
\value{
A matrix with the k-dimensional embedding.
}
\description{
Conveinience function to run isoMDS on a SingleCellExperiment object.
}
