% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{ispec}
\alias{ispec}
\title{Calculates the specific information which quantifies the information provided by one variable about a specific state of another variable [89, 90]}
\usage{
ispec(x, y, state)
}
\arguments{
\item{x}{The variable to quantify the information from}

\item{y}{The other variable}

\item{state}{The specific state of y}
}
\value{
The information provided by x about a specific state of y
}
\description{
Calculates the specific information which quantifies the information provided by one variable about a specific state of another variable [89, 90]
}

