% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{L2norm}
\alias{L2norm}
\title{Calculates the L2 norm of a numeric vector}
\usage{
L2norm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The L2 norm of x.
}
\description{
Calculates the L2 norm of a numeric vector
}

