% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_tests.R
\name{lrde_test}
\alias{lrde_test}
\title{Tests differential expression/accessibility using logistic regression}
\usage{
lrde_test(object, response, exprs_values = "logcounts",
  group_by = NULL, reference = NULL, features = NULL, ...)
}
\arguments{
\item{object}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{response}{Character vector indicating the response (class) variable. In case of length of one, it should determines the column name of the \code{\link[SummarizedExperiment]{colData}} slot.}

\item{exprs_values}{String indicating which assay contains the data that should be used to perform grouping and testing.}

\item{group_by}{Character vector indicating the grouping variable for expression values. In case of length of one, it should determines the column name of the \code{\link[SummarizedExperiment]{rowData}} slot.}

\item{reference}{The level of \code{response} that should be used as reference.}

\item{features}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to test. The default of NULL will use all features.}

\item{...}{Additional parameters passed to \code{\link[stats]{p.adjust}} for multiple testing correction.}
}
\value{
A tibble with the following columns:
  - group - the grouping variable
  - pvalue - the p-value from comparison of the two LR models
  - p.adjusted - the multiple testing corrected \code{pvalue}.
}
\description{
This function implements the test proposed by Ntranos et. al. (doi: 10.1038/s41592-018-0303-9) for \code{\link[SingleCellExperiment]{SingleCellExperiment}} objects.
Compared to the traditional approach of using cell labels as covariate, logistic regression is carried out to predict cell labels from
quantification of associated features. Those features can be transcripts or accessibility peaks associated with e.g. genes or regulatory regions.
}
