% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{map.init.local}
\alias{map.init.local}
\title{Initializes a SOM map with the first two PCA components in a localized fashion}
\usage{
map.init.local(data, resolution = 1)
}
\arguments{
\item{data}{The training data of the SOM}

\item{resolution}{A size factor increasing or decreasing the map resolution}
}
\value{
A list with initialization parameters
}
\description{
Initializes a SOM map with the first two PCA components in a localized fashion
}
