% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{mapinit}
\alias{mapinit}
\title{Initializes a SOM map with two PCA components}
\usage{
mapinit(datamat, prefactor = 1, toroidal = FALSE)
}
\arguments{
\item{datamat}{A numerical matrix or data.frame.}

\item{prefactor}{A factor.}

\item{toroidal}{If toroidal.}
}
\value{
A list with intialization parameters.
}
\description{
Initializes a SOM map with two PCA components
}

