% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{MeanBelowA}
\alias{MeanBelowA}
\title{MeanBelowA returns a filter function with bindings for A. This function evaluates to TRUE if the mean of the arguments elements are smaller than A.}
\usage{
MeanBelowA(A = 4, na.rm = TRUE)
}
\arguments{
\item{A}{The value you do not want to exceed.}

\item{na.rm}{Whether NAs should be removed.}
}
\value{
A function with bindings for A.
}
\description{
MeanBelowA returns a filter function with bindings for A. This function evaluates to TRUE if the mean of the arguments elements are smaller than A.
}
