% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{MeanOverA}
\alias{MeanOverA}
\title{MeanOverA returns a filter function with bindings for A. This function evaluates to TRUE if the mean of the arguments elements are greater than A.}
\usage{
MeanOverA(A = 4, na.rm = TRUE)
}
\arguments{
\item{A}{The value you want to exceed.}

\item{na.rm}{Whether NAs should be removed.}
}
\value{
A function with bindings for A.
}
\description{
MeanOverA returns a filter function with bindings for A. This function evaluates to TRUE if the mean of the arguments elements are greater than A.
}
