% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{n_filterfun}
\alias{n_filterfun}
\title{A custom \code{filterfun} that creates a n-FALSE exiting function from its input.}
\usage{
n_filterfun(n = 0, ...)
}
\arguments{
\item{n}{The number of tolerated FALSE exiting functions before the returned function returns FALSE. The behaviour of the default (n=0) resembles \code{\link[genefilter]{filterfun}}.}

\item{...}{A list of filter functions.}
}
\value{
A function.
}
\description{
A custom \code{filterfun} that creates a n-FALSE exiting function from its input.
}
\seealso{
\code{\link[genefilter]{filterfun}}.
}
