% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neuron_1k_v3}
\alias{neuron_1k_v3}
\title{1k Brain Cells from an E18 Mouse (v3 chemistry)}
\format{A SingleCellExperiment object with 31053 features measured across 1301 samples:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/3.0.0/neuron_1k_v3}
}
\usage{
neuron_1k_v3
}
\description{
Single Cell Gene Expression Dataset by Cell Ranger 3.0.0
Cells from a combined cortex, hippocampus and sub ventricular zone of an E18 mouse.
* 1,301 cells detected
* Sequenced on Illumina NovaSeq with approximately 71,000 reads per cell
* 28bp read1 (16bp Chromium barcode and 12bp UMI), 91bp read2 (transcript), and 8bp I7 sample barcode
* run with --expect-cells=1000
}
\keyword{datasets}
