% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{norm.theil.t}
\alias{norm.theil.t}
\title{Calculate the Theil's T statistic for each gene and normalize it to the dropout rate. Also calculates a p-value for the residual deviation from a normal distribution.}
\usage{
norm.theil.t(data, use.quantile = 0.05, ...)
}
\arguments{
\item{data}{A numerical matrix or data.frame.}

\item{use.quantile}{The quantile that has to be exceeded for fitting.}

\item{...}{other parameters passed to theils.t}
}
\value{
A list with the following components:
\itemize{
 \item d The linear models coeffictient,
 \item i The linear models intercept,
 \item theils.t A named vector containing Theils T statistic,
 \item dropin A named vector with 1 - dropout values,
 \item pval The p-value for the residuals deviation from the normal distribution.
 }
}
\description{
Calculate the Theil's T statistic for each gene and normalize it to the dropout rate. Also calculates a p-value for the residual deviation from a normal distribution.
}
