% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{normByBin}
\alias{normByBin}
\title{A generic function to normalize a statistic by binning another statistic}
\usage{
normByBin(stat, order_by = stat, func = mean, n_bins = 30,
  normalize = TRUE)
}
\arguments{
\item{stat}{A numeric vector.}

\item{order_by}{A numeric vector to bin by.}

\item{func}{A function to summarize the statistic within a bin.}

\item{n_bins}{The number of bins to create.}

\item{normalize}{Whether or not to normalize to the standard deviation.}
}
\value{
A vector with the difference to the summarized statistic
}
\description{
A generic function to normalize a statistic by binning another statistic
}
