% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.hvg}
\alias{plot.hvg}
\title{Draws a smoothScatter-Plot with hvg fit line and highlighted variable genes.}
\usage{
\method{plot}{hvg}(hvg.fit, n = 500)
}
\arguments{
\item{hvg.fit}{A list result from the hvg() function.}

\item{n}{Number of highly variable genes to highlight.}
}
\description{
Draws a smoothScatter-Plot with hvg fit line and highlighted variable genes.
}

