% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotComplexHeatmap}
\alias{plotComplexHeatmap}
\title{Creating a ComplexHeatmap from expression or reducedDim data of a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object}
\usage{
plotComplexHeatmap(object, features, exprs_values = "normcounts",
  use_dimred = NULL, split_by = NULL, rownames = NULL,
  scale = NULL, col = NULL, ...)
}
\arguments{
\item{object}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{features}{A character vector, vector of indices or a named list thereof. In case of a list, rows are split accordingly.}

\item{exprs_values}{String indicating which assay contains the data that should be used for plotting.}

\item{use_dimred}{A character string indicating which dimension reduction to use, or NULL.}

\item{split_by}{Character vector indicating by which columns should be split. In case of length of one, it determines a column name of the \code{\link[SummarizedExperiment]{colData}} slot.}

\item{rownames}{A string indicating a column name of \code{\link[SummarizedExperiment]{rowData}} slot, used as alternative rownames.}

\item{scale}{A logical, indicating whether data should be scaled before plotting.}

\item{col}{A vector of colors if the color mapping is discrete or a color mapping function if the matrix is continuous. See \code{\link[ComplexHeatmap]{Heatmap}}.}

\item{...}{Additional arguments passed on to the \code{\link[ComplexHeatmap]{Heatmap}} function.}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap-class}} object.
}
\description{
This function serve as a wrapper for the \code{\link[ComplexHeatmap]{Heatmap}} function and uses \code{features} to show expression values or columns from the \code{use_dimred} slot.
}
