% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSilhouette}
\alias{plotSilhouette}
\title{Plots a silhouette plot}
\usage{
plotSilhouette(object, use_dimred, clusters, ...)
}
\arguments{
\item{object}{A SinglecellExperiment object}

\item{use_dimred}{A character string indicating which dimension reduction to use}

\item{clusters}{A character string indicating which annotation to use as clusters}

\item{...}{Other arguments passed to other functions}
}
\value{
A ggplot object
}
\description{
Plots a silhouette plot
}
