% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{redundancy}
\alias{redundancy}
\title{Calculates the redundant contribution of x1 and x2 regarding the target variable z}
\usage{
redundancy(z, x1, x2)
}
\arguments{
\item{z}{The target varibale}

\item{x1}{The first source variable}

\item{x2}{The second source variable}
}
\value{
The amount of information provided by each variable within S about each state of the target Z
}
\description{
If we consider S = {X1 , X2 } and a target variable Z, the calculates the amount of information provided by each variable within S about each state of the target Z.
}

