% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SCTNormalize}
\alias{SCTNormalize}
\title{SCT Normalize Seurat Object}
\usage{
SCTNormalize(seurat_object, vars.to.regress = NULL, n.features = 3000)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell data.}

\item{vars.to.regress}{A character vector specifying the variables to regress out during normalization.}

\item{n.features}{An integer specifying the number of variable features to select.}
}
\value{
A normalized Seurat object.
}
\description{
This function performs SCT normalization on a Seurat object.
}
