% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{sigmoid}
\alias{sigmoid}
\title{Returns the value of the sigmoid function
1 - (1 / (1 + exp(a + (b*x)))}
\usage{
sigmoid(x, a, b)
}
\arguments{
\item{x}{Parameter x of the function}

\item{a}{Parameter a of the function}

\item{b}{Parameter b of the function}
}
\value{
The evaluated function value.
}
\description{
Returns the value of the sigmoid function
1 - (1 / (1 + exp(a + (b*x)))
}
