% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulatescSeq}
\alias{simulatescSeq}
\title{Simulates an single-cell RNA-seq experiment}
\usage{
simulatescSeq(genes = 5500, cells = 150, beta.shape1 = 3,
  beta.shape2 = 5, logis.loc = 1.2, logis.scale = 1.5)
}
\arguments{
\item{genes}{Number of genes in the experiment}

\item{cells}{Number of cells in the experiment}

\item{beta.shape1, beta.shape2}{Shape parameter for a beta distribution, from which burst size and burst frequency are estimated}

\item{logis.loc}{Location parameter of the bivariate logistic distribution}

\item{logis.scale}{Scale parameter of the bivariate logistic distribution}
}
\value{
An FPM expression matrix
}
\description{
Simulates an single-cell RNA-seq experiment
}
