% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{substitution_set}
\alias{substitution_set}
\title{Generate an substitution set of all strings that are \code{nedit} edits away from \code{str} under a given alphabet.}
\usage{
substitution_set(str, nedit, alphabet = c("A", "T", "G", "C"))
}
\arguments{
\item{str}{The string that is substituted.}

\item{nedit}{Maximum allowed edit distance.}

\item{alphabet}{Possible alphabet, from which substitutions are drawn.}
}
\value{
A vector of strings.
}
\description{
Generate an substitution set of all strings that are \code{nedit} edits away from \code{str} under a given alphabet.
}
