% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{summarizeSOM}
\alias{summarizeSOM}
\title{Calculates a summarized som for a list of components}
\usage{
summarizeSOM(som, codes = NULL, summarize.fun = mean, ...)
}
\arguments{
\item{som}{The SOM object}

\item{codes}{The components that should be sumarized}

\item{summarize.fun}{The function that is used to summarize each SOM node}

\item{...}{Additional parameters passed to summarize.fun

If \code{codes} is a list of component numbers, the summarization is done sequentially for each list.}
}
\value{
A summarized SOM codes matrix. Each column corresponds to a list element from the \code{codes} input.
}
\description{
Calculates a summarized som for a list of components
}

