% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{svd}
\alias{svd}
\alias{randomized_svd}
\alias{approximate_svd}
\title{Approximate or randomized singular value decomposition}
\usage{
randomized_svd(object, exprs_values, n_dims, features = NULL,
  skip = NULL, seed = NULL, ...)

approximate_svd(object, exprs_values, n_dims, features = NULL,
  skip = NULL, seed = NULL, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used to perform SVD.}

\item{n_dims}{The number of approximate singular values to calculate.}

\item{features}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use for SVD. The default of NULL will use all features.}

\item{skip}{A numeric vector indicating which singular values to set to zero (and remove).}

\item{seed}{A numeric seed to initialize the random number generator.}

\item{...}{Additional arguments passed on to \code{\link[rsvd]{rsvd}} or \code{\link[irlba]{irlba}}.}
}
\value{
A matrix of dimension \code{ncol(object)} x \code{dims(object) - length(skip)}.
}
\description{
Approximate SVD is performed using the implementation from \code{\link[irlba]{irlba}}, randomized SVD is performed using the implementation from \code{\link[rsvd]{rsvd}}.
}
