% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{tf_idf}
\alias{tf_idf}
\title{Calculates the term frequency-inverse document frequency.}
\usage{
tf_idf(object, exprs_values = "counts", features = NULL,
  tf.method = c("raw", "binary", "adjusted", "logscaled"),
  idf.weight = c("unary", "idf", "idf.smooth", "pidf"), slot = "tdifd")
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used to calculate TF-IDF.}

\item{features}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use for calculation The default of NULL will use all features.}

\item{tf.method}{A character vector indicating the variant for calculating the TF.}

\item{idf.weight}{A character vector indicating the weighting for IDF calculation.}

\item{slot}{Determines the name of the assay in \code{\link[SingleCellExperiment]{assays}} to use for weighted expression values.}
}
\value{
A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with modified \code{reducedDims} slot.
}
\description{
Calculates the term frequency-inverse document frequency.
}
