% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{theils.between}
\alias{theils.between}
\title{Calculates the Theils index between groups.}
\usage{
theils.between(x, groups)
}
\arguments{
\item{x}{A numeric vector}

\item{groups}{A grouping factor.}
}
\value{
The Theils between.
}
\description{
Calculates the Theils index between groups.
}
