% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{theils.t}
\alias{theils.t}
\title{Calculate Theil's T-Statistic for a vector}
\usage{
theils.t(x, winsorize = FALSE, treat.zeros = "exclude", epsilon = 0, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{winsorize}{If winsorization should be applied to x}

\item{treat.zeros}{How zeros should be treated. One of c("exclude","epsilon")}

\item{epsilon}{In case of treat.zeros equals to "epsilon", a small number is added to observations of x equal to zero.}

\item{...}{Parameters passed to winsorize, if applicable.}
}
\value{
Theil's T statistic for x
}
\description{
Calculate Theil's T-Statistic for a vector
}

