% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{theils.within}
\alias{theils.within}
\title{Calculates the Theils index within groups.}
\usage{
theils.within(x, groups)
}
\arguments{
\item{x}{A numeric vector}

\item{groups}{A grouping factor.}
}
\value{
The Theils within.
}
\description{
Calculates the Theils index within groups.
}
