% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{thin.counts}
\alias{thin.counts}
\title{Thinning of a count matrix}
\usage{
thin.counts(y, current.lib.sizes = colSums(y),
  target.lib.sizes = min(current.lib.sizes))
}
\arguments{
\item{y}{The count matrix}

\item{target.lib.sizes}{The target library sizes}

\item{current.lib.sized}{The current library sizes}
}
\value{
A list with two components:
  counts: The thinned count matrix
  lib.sizes: The new library sizes
}
\description{
Thinning of a count matrix
}
