% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{thin_counts}
\alias{thin_counts}
\title{Thinning of a expression matrix from a SingleCellExperiment object.}
\usage{
thin_counts(object, exprs_values,
  current_libsizes = scater::librarySizeFactors(object, exprs_values),
  target_libsizes = min(current_libsizes), return_as = "downsampled")
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used to perform thinning.}

\item{current_libsizes}{The current library sizes (can directly be calculated from \code{object}).}

\item{target_libsizes}{The target library sizes th thin expression to.}

\item{return_as}{String indicating the name of the assay which contains the thinned expression values.}
}
\value{
A SingleCellExperiment object with a new assay containing the thinned values.
}
\description{
Thinning of a expression matrix from a SingleCellExperiment object.
}
