% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{tSOM}
\alias{tSOM}
\title{Calculates a self-organizing map followed by t-SNE dimension reduction}
\usage{
tSOM(object, som.dots = list(), tsne.dots = list())
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{som.dots}{A named list with parameters passed on to \code{calcSOM}.}

\item{tsne.dots}{A named list with parameters passed on to \code{Rtsne::Rtsne}.}
}
\value{
A matrix with a two-dimensional embedding.
}
\description{
Calculates a self-organizing map followed by t-SNE dimension reduction
}
