% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visMarkerBar}
\alias{visMarkerBar}
\title{Creates a sorted bar plot of expression values.}
\usage{
visMarkerBar(object, exprs_values, feature, colour_by = NULL,
  threshold = 10^-2, decreasing = TRUE)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used to perform thinning.}

\item{feature}{A numerical index or character vector giving the rowname.}

\item{colour_by}{Character giving the name of a column in \code{colData} to color by.}

\item{threshold}{A lower bound for expression. Values below will be shown as 10^-2.}

\item{decreasing}{Whether the expression values should be sorted decreasing.}
}
\value{
A ggplot2 figure.
}
\description{
Creates a sorted bar plot of expression values.
}
