% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visSOM}
\alias{visSOM}
\title{Visualizes SOM components}
\usage{
visSOM(som, code = NULL, titles = NULL)
}
\arguments{
\item{som}{The self-organizing map object.}

\item{code}{The codes that should be displayed.}

\item{titles}{The title for each component.}
}
\value{
The matrix of color codes (invisible)
}
\description{
Visualizes SOM components
}

