% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vistSNE}
\alias{vistSNE}
\title{Visualizes tSNE maps with colored clusters by kmeans.}
\usage{
vistSNE(tsne, k = NULL, use.pal = "Dark2", add.center = T)
}
\arguments{
\item{tsne}{The data matrix from a tSNE dimension reduction call}

\item{k}{The number of clusters. If NULL, will perform Gap statistics.}

\item{use.pal}{The RColorBrewer palette name to use for cluster colors.}

\item{add.center}{Whether cluster centers should be added to the plot.}
}
\value{
The augmented input data matrix. Clusters are added in .cluster column.
}
\description{
Visualizes tSNE maps with colored clusters by kmeans.
}

