% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{VizSummarizedLoadings}
\alias{VizSummarizedLoadings}
\title{Visualize Summarized Loadings}
\usage{
VizSummarizedLoadings(seurat_object, dims = 15, genes = 20)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell data.}

\item{dims}{An integer specifying the number of principal components to consider.}

\item{genes}{An integer specifying the number of top genes to show in the plot.}
}
\value{
A bar plot showing the aggregated loadings across principal components.
}
\description{
This function visualizes the aggregated loadings across principal components.
}
