% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{winsorize}
\alias{winsorize}
\title{Winsorizes a vector to either a fraction or a absolute number}
\usage{
winsorize(x, fraction = 0.05, absolute = NULL, two.sided = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{fraction}{The fraction to calculate the limits}

\item{absolute}{The absolute number to calculate the limits}

\item{two.sided}{If winsorization should be done two-sided}
}
\value{
The winsorized vector
}
\description{
Winsorizes a vector to either a fraction or a absolute number
}

