% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{calcFNWeight}
\alias{calcFNWeight}
\title{Calculates weights/probabilities based on false-negative curves.}
\usage{
calcFNWeight(data, means = NULL, nq = 30, expression_cutoff = 10,
  supress.plot = T)
}
\arguments{
\item{data}{Expression data matrix}

\item{means}{A vector containing mean expression values, e.g. from a bulk measurement}

\item{nq}{Number of quantiles used for binning the means}

\item{expression_cutoff}{A numerical value}

\item{supress.plot}{Whether to supress plotting (default TRUE)}
}
\value{
A matrix with probabilites that a gene i in cell j is not expressed, given that it it not detected P(not expressed|not detected). The values are only meaningful if the expression of gene i in cell j is below the detection limit.
}
\description{
Calculates weights/probabilities based on false-negative curves.
}
