% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calcSOM}
\alias{calcSOM}
\title{Calculates a self-organizing map}
\usage{
calcSOM(data, sizemultiplier = 1, num_epochs = 200, train = NULL,
  seed = NULL, initrand = FALSE, init.map = NULL, intoroidal = FALSE)
}
\arguments{
\item{data}{A numerical matrix or data.frame.}

\item{sizemultiplier}{A size multiplier.}

\item{num_epochs}{How long the training should last.}

\item{train}{Row indices of training data.}

\item{seed}{A random seed (default NULL).}

\item{initrand}{Whether initialization should be random (not recommended).}

\item{intoroidal}{If a toroidal map should be created.}
}
\value{
A som object.
}
\description{
Calculates a self-organizing map
}

