% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{calcSOM}
\alias{calcSOM}
\title{Calculates a (weighted) self-organizing map}
\usage{
calcSOM(object, exprs_values = "norm_TPM", features = NULL, scale = T,
  num_epochs = 200, resolution = 1, seed = NULL, init = NULL,
  init.FUN = map.init, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{exprs_values}{String indicating which assay contains the data that should be used to perform SOM calculation.}

\item{features}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use for SOM calculation. The default of NULL will use all features.}

\item{scale}{Logical scalar indicating whether data should be scaled before SOM calculation.}

\item{num_epochs}{How long the training should last.}

\item{resolution}{A multiplier for the map size.}

\item{seed}{A numeric seed to initialize the random number generator.}

\item{init}{A list with initialization parameters.}

\item{init.FUN}{A function to initialize the map.}

\item{...}{Additional arguments passed to \code{kohonen::som}.}
}
\value{
A som object.
}
\description{
Calculates a (weighted) self-organizing map
}
