% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterSOM}
\alias{clusterSOM}
\title{Clusters a SOM into k regions determined by k-means}
\usage{
clusterSOM(codes, k = NULL, ...)
}
\arguments{
\item{codes}{The codes matrix from a SOM object.}

\item{k}{Number of clusters. If NULL, will perform Gap statistics.}

\item{...}{Additional parameters passed to hclust.}
}
\value{
A clustering of the SOM.
}
\description{
Clusters a SOM into k regions determined by k-means
}

