% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterSOM}
\alias{clusterSOM}
\title{Clusters a SOM into k regions determined by k-means}
\usage{
clusterSOM(som, kmax = NULL, dist.fun = median, even.layout = T,
  max.rounds = max(100, nrow(som$codes)/kmax), ...)
}
\arguments{
\item{som}{The SOM object.}

\item{kmax}{The number of clusters. If NULL, will determine k internally.}

\item{dist.fun}{Function to calculate neighborhood distances.}

\item{even.layout}{Whether the map layout is even (odd otherwise).}

\item{max.rounds}{Maximum number of cluster assignments.}

\item{...}{Additional parameters.}
}
\value{
A clustering of the SOM.
}
\description{
Clusters a SOM into k regions determined by k-means
}

