% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{colorAMap}
\alias{colorAMap}
\title{A function to color observations on a map based on values of a vector.}
\usage{
colorAMap(data, values, palette = RColorBrewer::brewer.pal(10, "Dark2"),
  title = NULL, outlier.col = "darkgrey", outlier.pch = 4, ...)
}
\arguments{
\item{data}{The data matrix with observation coordinates in the first two columns.}

\item{values}{A numerical vector on which the coloring is based, or a numeric or character of length 1 indicating the column of the data matrix from which to take the values.}

\item{palette}{The color palette}

\item{title}{The name of the data, shown in plot.main}

\item{outlier.col}{The color of outlier points (\code{NA} in \code{values})}

\item{outlier.pch}{The symbol for outlier points (\code{NA} in \code{values})}

\item{...}{Other parameters passed to plot}
}
\value{
Invisibly returns the color code for each observation.
}
\description{
A function to color observations on a map based on values of a vector.
}
