% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{colortSNE}
\alias{colortSNE}
\title{A function to color a tSNE map based on a vector.}
\usage{
colortSNE(tsne, data, data.name = NULL,
  palette = RColorBrewer::brewer.pal(10, "Dark2"), ...)
}
\arguments{
\item{tsne}{The data matrix from a tSNE dimension reduction call}

\item{data}{A numerical vector on which the coloring is based}

\item{data.name}{The name of the data, shown in plot.main}

\item{palette}{The color palette}

\item{...}{Other parameters passed to plot}
}
\description{
A function to color a tSNE map based on a vector.
}
