% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_index.R
\name{computeIC}
\alias{computeIC}
\title{Calculate IC values for a data matrix (genes-by-cells)}
\usage{
computeIC(data)
}
\arguments{
\item{data}{Data matrix}
}
\value{
The IC value
}
\description{
Calculate IC values for a data matrix (genes-by-cells)
}
