% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{hdbscan}
\alias{hdbscan}
\title{Conveinience function to perform HDBSCAN via reticulate.}
\usage{
hdbscan(object, use_dimred, min_samples = 7L, min_cluster_size = 9L,
  outlier = 0, seed = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{use_dimred}{A string or integer scalar indicating the reduced dimension result in \code{reducedDims(object)} to use as input.}

\item{min_samples}{Measure of how conservative the clustering should to be. The larger the value of \code{min_samples}, the more conservative the clustering – more points will be declared as noise, and clusters will be restricted to progressively more dense areas.}

\item{min_cluster_size}{The smallest size grouping that is considered a cluster.}

\item{outlier}{Determines how outliers are encoded in the resulting clustering.}

\item{seed}{A numeric seed to initialize the random number generator.}
}
\value{
A factor with the assigned cluster.
}
\description{
Performs HDBSCAN clustering on a SingleCellExperiment object.
The algorithm is explained in detail in http://joss.theoj.org/papers/10.21105/joss.00205 and needs the python package \code{hdbscan} installed.
}
