% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{norm.theil.t}
\alias{norm.theil.t}
\title{Calculate the normalized Theil's T statistic}
\usage{
norm.theil.t(data, ...)
}
\arguments{
\item{data}{A numerical matrix or data.frame.}

\item{window}{The window size to calculate the summary statistic in.}

\item{normalize}{Whether or not to normalize to the standard deviation.}

\item{func}{The function to calculate the summary statistic within windows.}
}
\value{
A vector with the normalize Theil's T statistic.
}
\description{
Calculate the normalized Theil's T statistic
}

