% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{normByWindow}
\alias{normByWindow}
\title{A generic function to normalize a statistic by a summary statistic across windows}
\usage{
normByWindow(stat, order_by = stat, func = stats::median, window = 100,
  normalize = TRUE)
}
\arguments{
\item{stat}{A numeric vector.}

\item{order_by}{A numeric vector to order stat by.}

\item{func}{A function to summarize the statistic within a window.}

\item{window}{The window size to calculate the summary statistic in.}

\item{normalize}{Whether or not to normalize to the standard deviation.}
}
\value{
A vector with the difference to the summarized statistic
}
\description{
A generic function to normalize a statistic by a summary statistic across windows
}
