% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{reduce_dimension}
\alias{reduce_dimension}
\title{Adds results from a particular dimension reduction technique to the object.}
\usage{
reduce_dimension(object, flavor = c("umap", "som", "som_tsne", "svd",
  "isomds", "rsvd"), slot = flavor, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{flavor}{Determines which dimension reduction technique to apply.}

\item{slot}{Determines which entry of the \code{reducedDims} slot to use for reduced embedding.}

\item{...}{Additional parameters passed to functions.}
}
\value{
A SingleCellExperiment object with modified \code{reducedDims} slot.
}
\description{
Adds results from a particular dimension reduction technique to the object.
}
