% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{scaterIsOutlier}
\alias{scaterIsOutlier}
\title{scaterIsOutlier returns a filter function with bindings for parameters for \code{scater::isOutlier}.}
\usage{
scaterIsOutlier(...)
}
\arguments{
\item{...}{Parameters that will be bound to the \code{scater::isOutlier} code.}
}
\value{
A function with bindings for ....
}
\description{
scaterIsOutlier returns a filter function with bindings for parameters for \code{scater::isOutlier}.
}
