% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{scaterIsOutlier}
\alias{scaterIsOutlier}
\title{A filter function with bindings for parameters for \code{\link[scater]{isOutlier}}.}
\usage{
scaterIsOutlier(...)
}
\arguments{
\item{...}{Parameters that will be bound to the \code{\link[scater]{isOutlier}} code.}
}
\value{
A function with ... bindings for \code{\link[scater]{isOutlier}}.
}
\description{
A filter function with bindings for parameters for \code{\link[scater]{isOutlier}}.
}
