% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visMarkerBar}
\alias{visMarkerBar}
\title{Creates a sorted bar plot.}
\usage{
visMarkerBar(data, marker, decreasing = TRUE, cell.names = NULL,
  cell.groups = NULL, threshold = 10^-2, y.unit = NULL)
}
\arguments{
\item{data}{The data matrix or data.frame.}

\item{marker}{A numerical index or rowname.}

\item{decreasing}{Whether the expression values should be sorted decreasing.}

\item{cell.names}{A vector of alternative cell names.}

\item{cell.groups}{A factor of cell groups. If specified, cells will be colored accordingly.}

\item{threshold}{A lower bound for expression. Values below will be shown as 10^-2.}

\item{y.unit}{A character giving the unit of expression.}
}
\value{
A ggplot2 figure.
}
\description{
Creates a sorted bar plot.
}

