% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visMarkerViolin}
\alias{visMarkerViolin}
\title{Creates a violin plot for selected genes}
\usage{
visMarkerViolin(data, gene.names = NULL, group = NULL)
}
\arguments{
\item{data}{The expression data (as data.frame)}

\item{gene.names}{A character vector of associated gene names (optional)}
}
\value{
A ggplot2 figure.
}
\description{
Creates a violin plot for selected genes
}

