% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visSOM}
\alias{visSOM}
\title{Visualizes SOM components}
\usage{
visSOM(som, codes = NULL, titles = NULL, what = c("code", "population",
  "aggregate", "overexpression", "underexpression"), aggregate.FUN = mean,
  scale = c("none", "minmax", "zscore"), palette = NULL)
}
\arguments{
\item{som}{The self-organizing map object.}

\item{titles}{The title for each component.}

\item{what}{A keyword for what to plot.}

\item{aggregate.FUN}{A function to use for aggregation of codes.}

\item{scale}{Which scaling should be applied.}

\item{palette}{The colour palette used for scaling. If NULL, will be chosen automatically.}

\item{code}{The codes that should be displayed.}
}
\value{
A ggplot2 object or a list of ggplot2 objects, in case of multiple maps.
}
\description{
Visualizes SOM components
}
